+++ src/callbacks.py
@@ -7,12 +7,12 @@ and `verify` (on_verify) operations.
 
 import argparse
 from constants import KSIGNER_VERSION
-from hashutils import open_and_hash_file, save_hashed_file
-from logutils import verbose_log
-from pemutils import create_public_key_certificate
-from qrutils import make_qr_code
-from signandverifyutils import verify
-from videoutils import scan_and_save_signature, scan_public_key
+from utils.hash import open_and_hash_file, save_hashed_file
+from utils.log import logger
+from utils.pem import create_public_key_certificate
+from utils.qr import make_qr_code
+from utils.signandverify import verify
+from utils.video import scan_and_save_signature, scan_public_key
 
 
 def on_version(parser: argparse.ArgumentParser):
@@ -52,33 +52,37 @@ def on_sign(parser: argparse.ArgumentParser):
     # If the signergn command was given
     if args.command == "sign" and args.file_to_sign is not None:
         # read file
-        data = open_and_hash_file(path=args.file_to_sign, verbose=args.verbose)
+
+        logger("DEBUG", f"opening {args.file_to_sign}")
+        data = open_and_hash_file(path=args.file_to_sign)
 
         # Saves a hash file
-        save_hashed_file(data=data, path=args.file_to_sign, verbose=args.verbose)
+        logger("DEBUG", f"saving {args.file_to_sign}.sha256.txt")
+        save_hashed_file(data=data, path=args.file_to_sign)
 
         # Shows some message
-        verbose_log("INFO", "To sign this file with Krux: ")
-        verbose_log("INFO", " (a) load a 24 words key;")
-        verbose_log("INFO", " (b) use the Sign->Message feature;")
-        verbose_log("INFO", " (c) and scan this QR code below.")
+        print("")
+        print("To sign this file with Krux: ")
+        print(" (a) load a 12/24 words key with or without password;")
+        print(" (b) use the Sign->Message feature;")
+        print(" (c) and scan this QR code below.")
+        print("")
 
         # Prints the QR code
+        logger("DEBUG", f"Adding (data={data})")
         __qrcode__ = make_qr_code(data=data, verbose=args.verbose)
-        print(f"\n{__qrcode__}")
+        print(f"{__qrcode__}")
 
         # Scans the signature QR code
         scan_and_save_signature(
             is_normalized=args.is_normalized,
             is_gray_scale=args.is_gray_scale,
-            verbose=args.verbose,
         )
 
         # Scans the public KeyboardInterruptardInterrupt
         pubkey = scan_public_key(
             is_normalized=args.is_normalized,
             is_gray_scale=args.is_gray_scale,
-            verbose=args.verbose,
         )
 
         # Create PEM data
@@ -88,7 +92,6 @@ def on_sign(parser: argparse.ArgumentParser):
             pubkey=pubkey,
             uncompressed=args.uncompressed_pub_key,
             owner=args.file_owner,
-            verbose=args.verbose,
         )
 
 
@@ -113,7 +116,6 @@ def on_verify(parser: argparse.ArgumentParser):
             filename=args.verify_file,
             pubkey=args.pub_file,
             sigfile=args.sig_file,
-            verbose=args.verbose,
         )
     # If command was not found
     else:
