+++ src/utils/qr.py
@@ -13,10 +13,6 @@ from io import StringIO
 #######################
 from qrcode import QRCode
 
-#################
-# Local libraries
-#################
-from logutils import verbose_log
 
 def make_qr_code(**kwargs) -> str:
     """
@@ -29,21 +25,10 @@ def make_qr_code(**kwargs) -> str:
             Apply verbose or not
     """
     qr_data = kwargs.get("data")
-    verbose = kwargs.get("verbose")
-
     qr_code = QRCode()
-
-    if verbose:
-        verbose_log("INFO", f"Adding (data={qr_data})")
-
     qr_code.add_data(qr_data)
-
-    if verbose:
-        verbose_log("INFO", "Converting data to ascii")
-
     qr_string = StringIO()
     qr_code.print_ascii(out=qr_string, invert=True)
-
     return qr_string.getvalue()
 
 
@@ -58,17 +43,7 @@ def make_qr_code_image(**kwargs) -> str:
             Apply verbose or not
     """
     qr_data = kwargs.get("data")
-    verbose = kwargs.get("verbose")
-
     qr_code = QRCode()
-
-    if verbose:
-        verbose_log("INFO", f"Adding (data={qr_data})")
-
     qr_code.add_data(qr_data)
-
-    if verbose:
-        verbose_log("INFO", "Creating image")
-
     qr_image = qr_code.make_image()
     return qr_image
